- Another change for old compiler (redhate 4)

+++ cpp_test_suite/new_tests/cxx_database.cpp
@@ -155,7 +155,7 @@ public:
 
 		TS_ASSERT_THROWS_NOTHING(db->get_class_pipe_property("MyStepperMotor",get_db_data));
 
-		int nb_prop;
+		short nb_prop;
 		get_db_data[0] >> nb_prop;
 		TS_ASSERT(nb_prop == 2);
